# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: null
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
  /api/v1/spaces:
    get:
      summary: Список
      operationId: ''
      description: ''
      parameters:
        -
          in: query
          name: search
          description: 'Поисковой запрос.'
          example: 'Моя доска'
          required: false
          schema:
            type: string
            description: 'Поисковой запрос.'
            example: 'Моя доска'
            nullable: true
        -
          in: query
          name: limit
          description: 'Кол-во элементов на странице.'
          example: 10
          required: false
          schema:
            type: integer
            description: 'Кол-во элементов на странице.'
            example: 10
            nullable: true
      responses:
        200:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: ''
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: forbidden
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'В доступе отказано'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'В доступе отказано'
        422:
          description: 'validation failed'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Errors message'
                  errors:
                    field:
                      - 'Error text'
                properties:
                  message:
                    type: string
                    example: 'Errors message'
                  errors:
                    type: object
                    properties:
                      field:
                        type: array
                        example:
                          - 'Error text'
                        items:
                          type: string
      tags:
        - 'API V1'
    post:
      summary: Создание
      operationId: ''
      description: ''
      parameters: []
      responses:
        201:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: ''
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: forbidden
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'В доступе отказано'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'В доступе отказано'
        422:
          description: 'validation failed'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Errors message'
                  errors:
                    field:
                      - 'Error text'
                properties:
                  message:
                    type: string
                    example: 'Errors message'
                  errors:
                    type: object
                    properties:
                      field:
                        type: array
                        example:
                          - 'Error text'
                        items:
                          type: string
        500:
          description: 'went wrong'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Что-то пошло не так.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Что-то пошло не так.'
      tags:
        - 'API V1'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_id:
                  type: string
                  description: 'ID-владельца. validation.uuid.'
                  example: null
                  nullable: false
                name:
                  type: string
                  description: 'Название пространства. validation.max.'
                  example: Название
                  nullable: false
                description:
                  type: string
                  description: 'Описание пространства. validation.max.'
                  example: 'Описание пространства'
                  nullable: true
                tariff:
                  type: string
                  description: Тариф.
                  example: 'Хз что тут должно быть'
                  nullable: false
              required:
                - owner_id
                - name
                - tariff
  '/api/v1/spaces/{modelId}':
    get:
      summary: Детали
      operationId: ''
      description: ''
      parameters: []
      responses:
        200:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: ''
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: forbidden
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'В доступе отказано'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'В доступе отказано'
        404:
          description: 'not found'
          content:
            text/plain:
              schema:
                type: string
                example: ''
      tags:
        - 'API V1'
    post:
      summary: Обновление
      operationId: ''
      description: ''
      parameters: []
      responses:
        200:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: ''
        400:
          description: 'bad request'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Проверьте правильность данных и повторите попытку'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Проверьте правильность данных и повторите попытку'
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: forbidden
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'В доступе отказано'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'В доступе отказано'
        404:
          description: 'not found'
          content:
            text/plain:
              schema:
                type: string
                example: ''
        422:
          description: 'validation failed'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Errors message'
                  errors:
                    field:
                      - 'Error text'
                properties:
                  message:
                    type: string
                    example: 'Errors message'
                  errors:
                    type: object
                    properties:
                      field:
                        type: array
                        example:
                          - 'Error text'
                        items:
                          type: string
        500:
          description: 'went wrong'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Что-то пошло не так.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Что-то пошло не так.'
      tags:
        - 'API V1'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Название пространства. validation.max.'
                  example: Название
                  nullable: true
                description:
                  type: string
                  description: 'Описание пространства. validation.max.'
                  example: 'Описание пространства'
                  nullable: true
                tariff:
                  type: string
                  description: Тариф.
                  example: 'Хз что тут должно быть'
                  nullable: true
    delete:
      summary: Удаление
      operationId: ''
      description: ''
      parameters: []
      responses:
        200:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: ''
        400:
          description: 'bad request'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Проверьте правильность данных и повторите попытку'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Проверьте правильность данных и повторите попытку'
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: forbidden
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'В доступе отказано'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'В доступе отказано'
        404:
          description: 'not found'
          content:
            text/plain:
              schema:
                type: string
                example: ''
      tags:
        - 'API V1'
    parameters:
      -
        in: path
        name: modelId
        description: ID-пространства
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        required: true
        schema:
          type: uuid
  /api/registration/verify:
    post:
      summary: ''
      operationId: postApiRegistrationVerify
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: validation.email.
                  example: gbailey@example.net
                  nullable: false
                verify_code:
                  type: integer
                  description: ''
                  example: 16
                  nullable: false
              required:
                - email
                - verify_code
      security: []
  /api/authorization:
    post:
      summary: ''
      operationId: postApiAuthorization
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: validation.email.
                  example: gbailey@example.net
                  nullable: false
                password:
                  type: string
                  description: ''
                  example: architecto
                  nullable: false
                remember_me:
                  type: boolean
                  description: ''
                  example: false
                  nullable: false
              required:
                - email
                - password
                - remember_me
      security: []
  /api/password/reset/email:
    post:
      summary: ''
      operationId: postApiPasswordResetEmail
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: validation.email.
                  example: gbailey@example.net
                  nullable: false
              required:
                - email
      security: []
  /api/password/reset/verify:
    post:
      summary: ''
      operationId: postApiPasswordResetVerify
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: validation.email.
                  example: gbailey@example.net
                  nullable: false
                verify_code:
                  type: integer
                  description: ''
                  example: 16
                  nullable: false
              required:
                - email
                - verify_code
      security: []
  /api/verify/code/resend:
    post:
      summary: ''
      operationId: postApiVerifyCodeResend
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: validation.email.
                  example: gbailey@example.net
                  nullable: false
              required:
                - email
      security: []
...
  /api/spaces/roles:
    post:
      summary: ''
      operationId: postApiSpacesRoles
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                space_id:
                  type: string
                  description: validation.uuid.
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: false
                name:
                  type: string
                  description: validation.max.
                  example: g
                  nullable: false
                description:
                  type: string
                  description: validation.max.
                  example: 'Quos velit et fugiat sunt nihil accusantium harum.'
                  nullable: true
                permissions:
                  type: object
                  description: ''
                  example: []
                  nullable: false
                  properties:
                    finance_access:
                      type: boolean
                      description: ''
                      example: false
                      nullable: false
                    projects_access:
                      type: boolean
                      description: ''
                      example: false
                      nullable: false
                    team_access:
                      type: boolean
                      description: ''
                      example: true
                      nullable: false
                    full_access:
                      type: boolean
                      description: ''
                      example: false
                      nullable: false
                  required:
                    - finance_access
                    - projects_access
                    - team_access
                    - full_access
              required:
                - space_id
                - name
                - permissions
      security: []
  '/api/spaces/roles/{id}':
    put:
      summary: ''
      operationId: putApiSpacesRolesId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: string
                  description: validation.uuid.
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: false
                name:
                  type: string
                  description: validation.max.
                  example: g
                  nullable: false
                description:
                  type: string
                  description: validation.max.
                  example: 'Quos velit et fugiat sunt nihil accusantium harum.'
                  nullable: false
                permissions:
                  type: object
                  description: ''
                  example: []
                  nullable: false
                  properties:
                    finance_access:
                      type: boolean
                      description: ''
                      example: false
                      nullable: false
                    projects_access:
                      type: boolean
                      description: ''
                      example: false
                      nullable: false
                    team_access:
                      type: boolean
                      description: ''
                      example: true
                      nullable: false
                    full_access:
                      type: boolean
                      description: ''
                      example: false
                      nullable: false
                  required:
                    - finance_access
                    - projects_access
                    - team_access
                    - full_access
              required:
                - name
                - description
                - permissions
      security: []
    delete:
      summary: ''
      operationId: deleteApiSpacesRolesId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: string
                  description: validation.uuid.
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: false
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the role.'
        example: architecto
        required: true
        schema:
          type: string
  /api/spaces/users:
    post:
      summary: ''
      operationId: postApiSpacesUsers
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                space_id:
                  type: string
                  description: validation.uuid.
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: false
                email:
                  type: string
                  description: validation.email.
                  example: ashly64@example.com
                  nullable: false
                role_id:
                  type: string
                  description: validation.uuid.
                  example: 21c4122b-d554-3723-966c-6d723ea5293f
                  nullable: false
              required:
                - space_id
                - email
                - role_id
      security: []
  '/api/spaces/users/{id}':
    delete:
      summary: ''
      operationId: deleteApiSpacesUsersId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                space_user_id:
                  type: string
                  description: validation.uuid.
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: false
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: architecto
        required: true
        schema:
          type: string
  /api/spaces:
    post:
      summary: ''
      operationId: postApiSpaces
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: validation.max.
                  example: b
                  nullable: false
                description:
                  type: string
                  description: validation.max.
                  example: 'Et animi quos velit et fugiat.'
                  nullable: true
                avatar:
                  type: string
                  description: 'Must be a valid URL.'
                  example: 'http://www.dach.com/mollitia-modi-deserunt-aut-ab-provident-perspiciatis-quo.html'
                  nullable: true
                admin_id:
                  type: string
                  description: ''
                  example: architecto
                  nullable: false
                tariff:
                  type: string
                  description: ''
                  example: Free
                  nullable: false
                  enum:
                    - Free
                    - Pro
                    - Enterprise
              required:
                - name
                - admin_id
                - tariff
      security: []
    get:
      summary: ''
      operationId: getApiSpaces
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Для использования данного функционала пользователь должен быть аутентифицированным'
                properties:
                  message:
                    type: string
                    example: 'Для использования данного функционала пользователь должен быть аутентифицированным'
      tags:
        - Endpoints
      security: []
...
  '/api/spaces/{id}/invite':
    post:
      summary: ''
      operationId: postApiSpacesIdInvite
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: validation.uuid.
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: false
                email:
                  type: string
                  description: validation.email.
                  example: ashly64@example.com
                  nullable: false
                role:
                  type: string
                  description: ''
                  example: architecto
                  nullable: false
              required:
                - email
                - role
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the space.'
        example: architecto
        required: true
        schema:
          type: string
...